refactor(domain): 중복 제거를 위한 객체를 추상화하여 상위 AbstractFirestore를 FirestoreFetcher로 변경 후 기능 이관
