--- Domains/hooks/Firestore/AbstractFirestore.ts
@@ -1,6 +1,0 @@
-abstract class AbstractFirestore {
-  constructor() {}
-  fetchData() {}
-}
-
-export default AbstractFirestore;
